# LPS-54590 inline permission performance fix . not final only for comments