--- a/.ivy/cache/com.liferay/com.liferay.source.formatter/ivy-1.0.11.xml.original
+++ b/.ivy/cache/com.liferay/com.liferay.source.formatter/ivy-1.0.11.xml.original
@@ -29,7 +29,7 @@
 		<developerConnection>scm:git:git@github.com:liferay/liferay-portal.git</developerConnection>
 		<url>http://github.com/liferay/liferay-portal</url>
 	</scm>
-	<version>1.0.10</version>
+	<version>1.0.11</version>
  <dependencies>
   <dependency>
    <groupId>org.apache.ant</groupId>