--- a/.ivy/cache/com.liferay/com.liferay.source.formatter/ivydata-1.0.11.properties
+++ b/.ivy/cache/com.liferay/com.liferay.source.formatter/ivydata-1.0.11.properties
@@ -1,0 +1,15 @@
+#ivy cached data file for com.liferay#com.liferay.source.formatter;1.0.11
+artifact.resolver=default
+artifact\:com.liferay.source.formatter\#jar\#jar\#-404634713.exists=true
+artifact\:com.liferay.source.formatter\#jar\#jar\#-404634713.is-local=true
+artifact\:com.liferay.source.formatter\#jar\#jar\#-404634713.location=file\:/Users/brian/.m2/repository/com/liferay/com.liferay.source.formatter/1.0.11/com.liferay.source.formatter-1.0.11.jar
+artifact\:com.liferay.source.formatter\#jar\#jar\#-404634713.original=artifact\:com.liferay.source.formatter\#jar\#jar\#-404634713
+artifact\:com.liferay.source.formatter\#pom.original\#pom\#1247927949.exists=true
+artifact\:com.liferay.source.formatter\#pom.original\#pom\#1247927949.is-local=true
+artifact\:com.liferay.source.formatter\#pom.original\#pom\#1247927949.location=file\:/Users/brian/.m2/repository/com/liferay/com.liferay.source.formatter/1.0.11/com.liferay.source.formatter-1.0.11.pom
+artifact\:com.liferay.source.formatter\#pom.original\#pom\#1247927949.original=artifact\:com.liferay.source.formatter\#pom.original\#pom\#1247927949
+artifact\:ivy\#ivy\#xml\#1979269432.exists=true
+artifact\:ivy\#ivy\#xml\#1979269432.is-local=true
+artifact\:ivy\#ivy\#xml\#1979269432.location=file\:/Users/brian/.m2/repository/com/liferay/com.liferay.source.formatter/1.0.11/com.liferay.source.formatter-1.0.11.pom
+artifact\:ivy\#ivy\#xml\#1979269432.original=artifact\:com.liferay.source.formatter\#pom.original\#pom\#1247927949
+resolver=local-m2