--- a/.ivy/cache/org.eclipse.equinox.bundles/rt.equinox.bundles/ivydata-4.5.0-SNAPSHOT.properties
+++ b/.ivy/cache/org.eclipse.equinox.bundles/rt.equinox.bundles/ivydata-4.5.0-SNAPSHOT.properties
@@ -2,10 +2,10 @@
 artifact.resolver=liferay-public
 artifact\:ivy\#ivy\#xml\#699062129.exists=true
 artifact\:ivy\#ivy\#xml\#699062129.is-local=false
-artifact\:ivy\#ivy\#xml\#699062129.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/bundles/rt.equinox.bundles/4.5.0-SNAPSHOT/rt.equinox.bundles-4.5.0-20150308.215729-3.pom
+artifact\:ivy\#ivy\#xml\#699062129.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/bundles/rt.equinox.bundles/4.5.0-SNAPSHOT/rt.equinox.bundles-4.5.0-20150611.165753-4.pom
 artifact\:ivy\#ivy\#xml\#699062129.original=artifact\:rt.equinox.bundles\#pom.original\#pom\#926026400
 artifact\:rt.equinox.bundles\#pom.original\#pom\#926026400.exists=true
 artifact\:rt.equinox.bundles\#pom.original\#pom\#926026400.is-local=false
-artifact\:rt.equinox.bundles\#pom.original\#pom\#926026400.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/bundles/rt.equinox.bundles/4.5.0-SNAPSHOT/rt.equinox.bundles-4.5.0-20150308.215729-3.pom
+artifact\:rt.equinox.bundles\#pom.original\#pom\#926026400.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/bundles/rt.equinox.bundles/4.5.0-SNAPSHOT/rt.equinox.bundles-4.5.0-20150611.165753-4.pom
 artifact\:rt.equinox.bundles\#pom.original\#pom\#926026400.original=artifact\:rt.equinox.bundles\#pom.original\#pom\#926026400
 resolver=liferay-public