--- a/.ivy/cache/org.eclipse.equinox.framework/rt.equinox.framework/ivydata-4.5.0-SNAPSHOT.properties
+++ b/.ivy/cache/org.eclipse.equinox.framework/rt.equinox.framework/ivydata-4.5.0-SNAPSHOT.properties
@@ -2,10 +2,10 @@
 artifact.resolver=liferay-public
 artifact\:ivy\#ivy\#xml\#-551968979.exists=true
 artifact\:ivy\#ivy\#xml\#-551968979.is-local=false
-artifact\:ivy\#ivy\#xml\#-551968979.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/framework/rt.equinox.framework/4.5.0-SNAPSHOT/rt.equinox.framework-4.5.0-20150308.215709-2.pom
+artifact\:ivy\#ivy\#xml\#-551968979.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/framework/rt.equinox.framework/4.5.0-SNAPSHOT/rt.equinox.framework-4.5.0-20150611.165811-3.pom
 artifact\:ivy\#ivy\#xml\#-551968979.original=artifact\:rt.equinox.framework\#pom.original\#pom\#-371860895
 artifact\:rt.equinox.framework\#pom.original\#pom\#-371860895.exists=true
 artifact\:rt.equinox.framework\#pom.original\#pom\#-371860895.is-local=false
-artifact\:rt.equinox.framework\#pom.original\#pom\#-371860895.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/framework/rt.equinox.framework/4.5.0-SNAPSHOT/rt.equinox.framework-4.5.0-20150308.215709-2.pom
+artifact\:rt.equinox.framework\#pom.original\#pom\#-371860895.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/framework/rt.equinox.framework/4.5.0-SNAPSHOT/rt.equinox.framework-4.5.0-20150611.165811-3.pom
 artifact\:rt.equinox.framework\#pom.original\#pom\#-371860895.original=artifact\:rt.equinox.framework\#pom.original\#pom\#-371860895
 resolver=liferay-public