--- a/.ivy/cache/org.eclipse.equinox/org.eclipse.equinox.http.servlet/ivydata-1.2.0-SNAPSHOT.properties
+++ b/.ivy/cache/org.eclipse.equinox/org.eclipse.equinox.http.servlet/ivydata-1.2.0-SNAPSHOT.properties
@@ -2,14 +2,14 @@
 artifact.resolver=default
 artifact\:ivy\#ivy\#xml\#-450392173.exists=true
 artifact\:ivy\#ivy\#xml\#-450392173.is-local=false
-artifact\:ivy\#ivy\#xml\#-450392173.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/org.eclipse.equinox.http.servlet/1.2.0-SNAPSHOT/org.eclipse.equinox.http.servlet-1.2.0-20150319.180852-15.pom
+artifact\:ivy\#ivy\#xml\#-450392173.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/org.eclipse.equinox.http.servlet/1.2.0-SNAPSHOT/org.eclipse.equinox.http.servlet-1.2.0-20150612.003607-17.pom
 artifact\:ivy\#ivy\#xml\#-450392173.original=artifact\:org.eclipse.equinox.http.servlet\#pom.original\#pom\#-2024071149
 artifact\:org.eclipse.equinox.http.servlet\#eclipse-plugin\#jar\#-2058684709.exists=true
 artifact\:org.eclipse.equinox.http.servlet\#eclipse-plugin\#jar\#-2058684709.is-local=false
-artifact\:org.eclipse.equinox.http.servlet\#eclipse-plugin\#jar\#-2058684709.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/org.eclipse.equinox.http.servlet/1.2.0-SNAPSHOT/org.eclipse.equinox.http.servlet-1.2.0-20150319.180852-15.jar
+artifact\:org.eclipse.equinox.http.servlet\#eclipse-plugin\#jar\#-2058684709.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/org.eclipse.equinox.http.servlet/1.2.0-SNAPSHOT/org.eclipse.equinox.http.servlet-1.2.0-20150612.003607-17.jar
 artifact\:org.eclipse.equinox.http.servlet\#eclipse-plugin\#jar\#-2058684709.original=artifact\:org.eclipse.equinox.http.servlet\#eclipse-plugin\#jar\#-2058684709
 artifact\:org.eclipse.equinox.http.servlet\#pom.original\#pom\#-2024071149.exists=true
 artifact\:org.eclipse.equinox.http.servlet\#pom.original\#pom\#-2024071149.is-local=false
-artifact\:org.eclipse.equinox.http.servlet\#pom.original\#pom\#-2024071149.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/org.eclipse.equinox.http.servlet/1.2.0-SNAPSHOT/org.eclipse.equinox.http.servlet-1.2.0-20150319.180852-15.pom
+artifact\:org.eclipse.equinox.http.servlet\#pom.original\#pom\#-2024071149.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/equinox/org.eclipse.equinox.http.servlet/1.2.0-SNAPSHOT/org.eclipse.equinox.http.servlet-1.2.0-20150612.003607-17.pom
 artifact\:org.eclipse.equinox.http.servlet\#pom.original\#pom\#-2024071149.original=artifact\:org.eclipse.equinox.http.servlet\#pom.original\#pom\#-2024071149
 resolver=liferay-public