--- a/.ivy/cache/org.eclipse.osgi/org.eclipse.osgi.services/ivydata-3.5.0-SNAPSHOT.properties
+++ b/.ivy/cache/org.eclipse.osgi/org.eclipse.osgi.services/ivydata-3.5.0-SNAPSHOT.properties
@@ -2,14 +2,14 @@
 artifact.resolver=default
 artifact\:ivy\#ivy\#xml\#-684849776.exists=true
 artifact\:ivy\#ivy\#xml\#-684849776.is-local=false
-artifact\:ivy\#ivy\#xml\#-684849776.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi.services/3.5.0-SNAPSHOT/org.eclipse.osgi.services-3.5.0-20150308.215518-2.pom
+artifact\:ivy\#ivy\#xml\#-684849776.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi.services/3.5.0-SNAPSHOT/org.eclipse.osgi.services-3.5.0-20150611.165350-3.pom
 artifact\:ivy\#ivy\#xml\#-684849776.original=artifact\:org.eclipse.osgi.services\#pom.original\#pom\#1691287438
 artifact\:org.eclipse.osgi.services\#eclipse-plugin\#jar\#1656673878.exists=true
 artifact\:org.eclipse.osgi.services\#eclipse-plugin\#jar\#1656673878.is-local=false
-artifact\:org.eclipse.osgi.services\#eclipse-plugin\#jar\#1656673878.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi.services/3.5.0-SNAPSHOT/org.eclipse.osgi.services-3.5.0-20150308.215518-2.jar
+artifact\:org.eclipse.osgi.services\#eclipse-plugin\#jar\#1656673878.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi.services/3.5.0-SNAPSHOT/org.eclipse.osgi.services-3.5.0-20150611.165350-3.jar
 artifact\:org.eclipse.osgi.services\#eclipse-plugin\#jar\#1656673878.original=artifact\:org.eclipse.osgi.services\#eclipse-plugin\#jar\#1656673878
 artifact\:org.eclipse.osgi.services\#pom.original\#pom\#1691287438.exists=true
 artifact\:org.eclipse.osgi.services\#pom.original\#pom\#1691287438.is-local=false
-artifact\:org.eclipse.osgi.services\#pom.original\#pom\#1691287438.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi.services/3.5.0-SNAPSHOT/org.eclipse.osgi.services-3.5.0-20150308.215518-2.pom
+artifact\:org.eclipse.osgi.services\#pom.original\#pom\#1691287438.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi.services/3.5.0-SNAPSHOT/org.eclipse.osgi.services-3.5.0-20150611.165350-3.pom
 artifact\:org.eclipse.osgi.services\#pom.original\#pom\#1691287438.original=artifact\:org.eclipse.osgi.services\#pom.original\#pom\#1691287438
 resolver=liferay-public