--- a/.ivy/cache/org.eclipse.osgi/org.eclipse.osgi/ivy-3.10.100-SNAPSHOT.xml.original
+++ b/.ivy/cache/org.eclipse.osgi/org.eclipse.osgi/ivy-3.10.100-SNAPSHOT.xml.original
@@ -30,7 +30,7 @@
         <version>${tycho.version}</version>
         <configuration>
           <compilerArgs>
-            <arg>-warn:-raw,unchecked,hiding,unused</arg>
+            <arg>-warn:-raw,unchecked,hiding,unused,warningToken</arg>
           </compilerArgs>
         </configuration>
       </plugin>