--- a/.ivy/cache/org.eclipse.osgi/org.eclipse.osgi/ivydata-3.10.100-SNAPSHOT.properties
+++ b/.ivy/cache/org.eclipse.osgi/org.eclipse.osgi/ivydata-3.10.100-SNAPSHOT.properties
@@ -2,14 +2,14 @@
 artifact.resolver=default
 artifact\:ivy\#ivy\#xml\#321541239.exists=true
 artifact\:ivy\#ivy\#xml\#321541239.is-local=false
-artifact\:ivy\#ivy\#xml\#321541239.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi/3.10.100-SNAPSHOT/org.eclipse.osgi-3.10.100-20150308.215431-2.pom
+artifact\:ivy\#ivy\#xml\#321541239.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi/3.10.100-SNAPSHOT/org.eclipse.osgi-3.10.100-20150611.165302-3.pom
 artifact\:ivy\#ivy\#xml\#321541239.original=artifact\:org.eclipse.osgi\#pom.original\#pom\#1393373799
 artifact\:org.eclipse.osgi\#eclipse-plugin\#jar\#1358760239.exists=true
 artifact\:org.eclipse.osgi\#eclipse-plugin\#jar\#1358760239.is-local=false
-artifact\:org.eclipse.osgi\#eclipse-plugin\#jar\#1358760239.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi/3.10.100-SNAPSHOT/org.eclipse.osgi-3.10.100-20150308.215431-2.jar
+artifact\:org.eclipse.osgi\#eclipse-plugin\#jar\#1358760239.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi/3.10.100-SNAPSHOT/org.eclipse.osgi-3.10.100-20150611.165302-3.jar
 artifact\:org.eclipse.osgi\#eclipse-plugin\#jar\#1358760239.original=artifact\:org.eclipse.osgi\#eclipse-plugin\#jar\#1358760239
 artifact\:org.eclipse.osgi\#pom.original\#pom\#1393373799.exists=true
 artifact\:org.eclipse.osgi\#pom.original\#pom\#1393373799.is-local=false
-artifact\:org.eclipse.osgi\#pom.original\#pom\#1393373799.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi/3.10.100-SNAPSHOT/org.eclipse.osgi-3.10.100-20150308.215431-2.pom
+artifact\:org.eclipse.osgi\#pom.original\#pom\#1393373799.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/osgi/org.eclipse.osgi/3.10.100-SNAPSHOT/org.eclipse.osgi-3.10.100-20150611.165302-3.pom
 artifact\:org.eclipse.osgi\#pom.original\#pom\#1393373799.original=artifact\:org.eclipse.osgi\#pom.original\#pom\#1393373799
 resolver=liferay-public