--- a/.ivy/cache/org.eclipse/eclipse-platform-parent/ivydata-4.5.0-SNAPSHOT.properties
+++ b/.ivy/cache/org.eclipse/eclipse-platform-parent/ivydata-4.5.0-SNAPSHOT.properties
@@ -2,10 +2,10 @@
 artifact.resolver=liferay-public
 artifact\:eclipse-platform-parent\#pom.original\#pom\#-2141392584.exists=true
 artifact\:eclipse-platform-parent\#pom.original\#pom\#-2141392584.is-local=false
-artifact\:eclipse-platform-parent\#pom.original\#pom\#-2141392584.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/eclipse-platform-parent/4.5.0-SNAPSHOT/eclipse-platform-parent-4.5.0-20150308.215140-2.pom
+artifact\:eclipse-platform-parent\#pom.original\#pom\#-2141392584.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/eclipse-platform-parent/4.5.0-SNAPSHOT/eclipse-platform-parent-4.5.0-20150611.165155-3.pom
 artifact\:eclipse-platform-parent\#pom.original\#pom\#-2141392584.original=artifact\:eclipse-platform-parent\#pom.original\#pom\#-2141392584
 artifact\:ivy\#ivy\#xml\#1051329785.exists=true
 artifact\:ivy\#ivy\#xml\#1051329785.is-local=false
-artifact\:ivy\#ivy\#xml\#1051329785.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/eclipse-platform-parent/4.5.0-SNAPSHOT/eclipse-platform-parent-4.5.0-20150308.215140-2.pom
+artifact\:ivy\#ivy\#xml\#1051329785.location=http\://cdn.repository.liferay.com/nexus/content/groups/public/org/eclipse/eclipse-platform-parent/4.5.0-SNAPSHOT/eclipse-platform-parent-4.5.0-20150611.165155-3.pom
 artifact\:ivy\#ivy\#xml\#1051329785.original=artifact\:eclipse-platform-parent\#pom.original\#pom\#-2141392584
 resolver=liferay-public