--- a/modules/util/source-formatter/bnd.bnd
+++ b/modules/util/source-formatter/bnd.bnd
@@ -1,6 +1,6 @@
 Bundle-Name: Liferay Source Formatter
 Bundle-SymbolicName: com.liferay.source.formatter
-Bundle-Version: 1.0.11
+Bundle-Version: 1.0.12
 Import-Package:\
 	!javax.portlet,\
 	!javax.servlet.http,\