… .ivy/cache/com.liferay/com.liferay.source.formatter/ivy-1.0.11.xml
… .ivy/cache/com.liferay/com.liferay.source.formatter/ivydata-1.0.10.properties
… .ivy/cache/com.liferay/com.liferay.source.formatter/jars/com.liferay.source.formatter-1.0.11.jar
… .ivy/cache/org.eclipse.equinox.bundles/rt.equinox.bundles/ivy-4.5.0-SNAPSHOT.xml
… .ivy/cache/org.eclipse.equinox.framework/rt.equinox.framework/ivy-4.5.0-SNAPSHOT.xml
… .ivy/cache/org.eclipse.equinox/org.eclipse.equinox.http.servlet/eclipse-plugins/org.eclipse.equinox.http.servlet-1.2.0-SNAPSHOT.jar
… .ivy/cache/org.eclipse.equinox/org.eclipse.equinox.http.servlet/ivy-1.2.0-SNAPSHOT.xml
… .ivy/cache/org.eclipse.osgi/org.eclipse.osgi.services/eclipse-plugins/org.eclipse.osgi.services-3.5.0-SNAPSHOT.jar
… .ivy/cache/org.eclipse.osgi/org.eclipse.osgi.services/ivy-3.5.0-SNAPSHOT.xml
… .ivy/cache/org.eclipse.osgi/org.eclipse.osgi/eclipse-plugins/org.eclipse.osgi-3.10.100-SNAPSHOT.jar
… .ivy/cache/org.eclipse.osgi/org.eclipse.osgi/ivy-3.10.100-SNAPSHOT.xml
… .ivy/cache/org.eclipse/eclipse-platform-parent/ivy-4.5.0-SNAPSHOT.xml
… build-test.xml
… git-commit-plugins
… modules/util/source-formatter/src/com/liferay/source/formatter/BaseSourceProcessor.java
… modules/util/source-formatter/src/com/liferay/source/formatter/JSPSourceProcessor.java
… modules/util/source-formatter/src/com/liferay/source/formatter/JavaImportsFormatter.java
… modules/util/source-formatter/src/com/liferay/source/formatter/JavaSourceProcessor.java
… modules/util/source-formatter/src/com/liferay/source/formatter/JavaTerm.java
… modules/util/source-formatter/src/com/liferay/source/formatter/PropertiesSourceProcessor.java
… modules/util/source-formatter/src/com/liferay/source/formatter/XMLSourceProcessor.java
… modules/util/sync-engine/src/com/liferay/sync/engine/documentlibrary/event/GetSyncDLObjectUpdateEvent.java
… modules/util/sync-engine/src/com/liferay/sync/engine/documentlibrary/handler/GetSyncDLObjectUpdateHandler.java
… modules/util/sync-engine/src/com/liferay/sync/engine/filesystem/SyncWatchEventProcessor.java
… modules/util/sync-engine/src/com/liferay/sync/engine/util/FileUtil.java
… portal-impl/src/META-INF/portal-hbm.xml
… portal-impl/src/META-INF/portal-model-hints.xml
… portal-impl/src/com/liferay/portal/model/impl/ResourcePermissionCacheModel.java
… portal-impl/src/com/liferay/portal/model/impl/ResourcePermissionImpl.java
… portal-impl/src/com/liferay/portal/model/impl/ResourcePermissionModelImpl.java