Hi Ray,

I'm sending this for your to comment. It is not complete since it does not have upgrade part yet and I did not implement that since if we would do some more refactoring here for data structure.

Idea here is that we do not support inline permissions other types of primary keys than long. At portal we seem to use only long as primary ( or String at counter), but filterFindBy is not supported anyway with string key.

There is test application:

https://github.com/sammso/liferay-portal/tree/permission-tester w

Which personally was good practise for OSGi + ServiceBuilder experience, and with this app I was able to verify the performance difference.

This test contains 1000 sites with 4 teams each and every site has 1000 entities scoped.

Performance without fix:
filterFindByGroup(20177, 0, 20); time : 37643 ms
filterCountByGroup(20177); time : 37461 ms

With fix:
filterFindByGroup(20177, 0, 20); time : 63 ms
filterCountByGroup(20177); time : 20 ms

So you can see that performance gain is proper :).

So I would like to hear your comments around and we can discuss on Skype too.